ボスのたまとゴーストをOBBで　ガード状態解除わかるように　難易度上げるためにスト本に魂打たす　ばぐなおし

+++ SourceCode/gamesystem/character/GuardAction.h
@@ -29,7 +29,7 @@ private:
 
     float GuardEaseT;
     bool GuardStart = false;
-    int GuardTime;
+    float GuardTime;
 private:
     void GuardAreacreate();
 public:
